# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is matchmaker server.",
        "title": "Matchmaker API",
        "contact": {}
    },
    "host": "localhost:8900",
    "paths": {
        "/matchmaker/users": {
            "post": {
                "description": "Handles request to add a new user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "matching"
                ],
                "summary": "Adding a user.",
                "parameters": [
                    {
                        "description": "Player",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AddRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created Player",
                        "schema": {
                            "$ref": "#/definitions/models.Player"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AddRequest": {
            "type": "object",
            "properties": {
                "latency": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "skill": {
                    "type": "number"
                }
            }
        },
        "models.Player": {
            "type": "object",
            "properties": {
                "added": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "latency": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "skill": {
                    "type": "number"
                }
            }
        }
    }
}